backend-python/requirements.txt
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
defusedxml==0.7.1
h11==0.16.0
httpcore==1.0.9
httpie==3.2.4
httpx==0.28.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.5.0
numpy==2.3.0
pandas==2.3.0
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
tzdata==2025.2
urllib3==2.4.0
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn==0.29.0